requirements.txt
fastapi==0.68.1
kafka-python==2.0.2
prometheus-fastapi-instrumentator==0.7.1
sentry-sdk==1.4.3
pydantic==1.8.2
